# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "File Conversion Platform",
    "description": "Comprehensive file conversion platform supporting 100+ conversion types",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/archives/convert/{conversion_id}": {
      "post": {
        "tags": [
          "archives"
        ],
        "summary": "Convert Archive",
        "description": "Convert between archive formats.\n\nSupported conversions:\n- ZIP Γåö RAR\n- RAR ΓåÆ ZIP\n- 7Z ΓåÆ ZIP\n- TAR ΓåÆ ZIP\n- GZIP ΓåÆ ZIP\n- ZIP ΓåÆ TAR",
        "operationId": "convert_archive_api_archives_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_archive_api_archives_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai/convert/{conversion_id}": {
      "post": {
        "tags": [
          "ai-powered"
        ],
        "summary": "Convert Ai",
        "description": "Perform AI-powered conversions.\n\nSupported conversions:\n- Image ΓåÆ Text (OCR)\n- PDF ΓåÆ Searchable PDF (OCR)\n- PDF ΓåÆ Text (OCR)\n- Handwritten ΓåÆ Text (OCR)\n- Image/Document ΓåÆ Structured JSON\n\nNote: Requires Tesseract OCR to be installed on the system.",
        "operationId": "convert_ai_api_ai_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_ai_api_ai_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/upload": {
      "post": {
        "summary": "Upload File",
        "description": "Upload a file for conversion\n\n- **file**: File to upload\n- **target_format**: Target format for conversion (optional, can be specified later)",
        "operationId": "upload_file_api_upload_post",
        "parameters": [
          {
            "name": "target_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Target Format"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversions/{conversion_id}": {
      "get": {
        "summary": "Get Conversion",
        "description": "Get conversion status and details",
        "operationId": "get_conversion_api_conversions__conversion_id__get",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Conversion",
        "description": "Delete a conversion and its files",
        "operationId": "delete_conversion_api_conversions__conversion_id__delete",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversions": {
      "get": {
        "summary": "List Conversions",
        "description": "List recent conversions",
        "operationId": "list_conversions_api_conversions_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/download/{conversion_id}": {
      "get": {
        "summary": "Download File",
        "description": "Download converted file",
        "operationId": "download_file_api_download__conversion_id__get",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_convert_ai_api_ai_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_ai_api_ai_convert__conversion_id__post"
      },
      "Body_convert_archive_api_archives_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_archive_api_archives_convert__conversion_id__post"
      },
...
      "Body_upload_file_api_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_upload_post"
      },
      "ConversionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "target_format": {
            "type": "string",
            "title": "Target Format"
          },
          "file_size": {
            "type": "number",
            "title": "File Size"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "source_format",
          "target_format",
          "file_size",
          "status",
          "created_at"
        ],
        "title": "ConversionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}